+++ compiler/src/main/kotlin/dev/zacsweers/metro/compiler/ir/IrMetroContext.kt
@@ -15,11 +15,17 @@ import dev.zacsweers.metro.compiler.tracing.Tracer
 import dev.zacsweers.metro.compiler.tracing.tracer
 import java.io.File
 import java.nio.file.Path
+import java.time.LocalDateTime
+import java.time.format.DateTimeFormatterBuilder
+import java.time.temporal.ChronoField
+import kotlin.io.path.ExperimentalPathApi
 import kotlin.io.path.appendText
 import kotlin.io.path.createDirectories
 import kotlin.io.path.createFile
 import kotlin.io.path.createParentDirectories
 import kotlin.io.path.deleteIfExists
+import kotlin.io.path.deleteRecursively
+import kotlin.io.path.exists
 import kotlin.io.path.writeText
 import org.jetbrains.kotlin.backend.common.extensions.IrPluginContext
 import org.jetbrains.kotlin.cli.common.messages.CompilerMessageSeverity
@@ -132,6 +138,21 @@ internal interface IrMetroContext : IrPluginContext, CompatContext {
   }
 
   companion object {
+    private val reportsDirNameFormatter =
+      DateTimeFormatterBuilder()
+        .appendValue(ChronoField.YEAR, 4)
+        .appendValue(ChronoField.MONTH_OF_YEAR, 2)
+        .appendValue(ChronoField.DAY_OF_MONTH, 2)
+        .appendLiteral('T')
+        .appendValue(ChronoField.HOUR_OF_DAY, 2)
+        .appendValue(ChronoField.MINUTE_OF_HOUR, 2)
+        .optionalStart()
+        .appendValue(ChronoField.SECOND_OF_MINUTE, 2)
+        .optionalStart()
+        .appendLiteral('_')
+        .appendFraction(ChronoField.NANO_OF_SECOND, 1, 1, false)
+        .toFormatter()
+
     operator fun invoke(
       pluginContext: IrPluginContext,
       messageCollector: MessageCollector,
@@ -193,9 +214,20 @@ internal interface IrMetroContext : IrPluginContext, CompatContext {
 
       override val irTypeSystemContext: IrTypeSystemContext =
         IrTypeSystemContextImpl(pluginContext.irBuiltIns)
+
       private val loggerCache = mutableMapOf<MetroLogger.Type, MetroLogger>()
 
-      override val reportsDir: Path? by lazy { options.reportsDestination?.createDirectories() }
+      @OptIn(ExperimentalPathApi::class)
+      override val reportsDir: Path? by lazy {
+        options.reportsDestination
+          ?.resolve(reportsDirNameFormatter.format(LocalDateTime.now()))
+          ?.run {
+            if (exists()) {
+              deleteRecursively()
+            }
+            createDirectories()
+          }
+      }
 
       override val logFile: Path? by lazy {
         reportsDir?.let {