+++ gradle-plugin/src/functionalTest/kotlin/dev/zacsweers/metro/gradle/incremental/BaseIncrementalCompilationTest.kt
@@ -11,22 +11,31 @@ import org.intellij.lang.annotations.Language
 
 abstract class BaseIncrementalCompilationTest {
 
-  protected val GradleProject.buildDir: File
+  protected val GradleProject.asMetroProject: MetroGradleProject
+    get() = MetroGradleProject(rootDir)
+
+  protected fun GradleProject.metroProject(path: String): MetroGradleProject {
+    return MetroGradleProject(rootDir.resolve(path))
+  }
+
+  @JvmInline protected value class MetroGradleProject(val rootDir: File)
+
+  protected val MetroGradleProject.buildDir: File
     get() = rootDir.resolve("build")
 
-  protected val GradleProject.metroDir: File
+  protected val MetroGradleProject.metroDir: File
     get() = buildDir.resolve("metro")
 
-  protected fun GradleProject.reports(compilation: String): Reports =
-    metroDir.resolve(compilation).let(::Reports)
+  protected fun MetroGradleProject.reports(compilation: String): Reports =
+    metroDir.resolve(compilation).listFiles().maxByOrNull { it.lastModified() }!!.let(::Reports)
 
-  protected val GradleProject.mainReports: Reports
+  protected val MetroGradleProject.mainReports: Reports
     get() = reports("main")
 
-  protected val GradleProject.appGraphReports: GraphReports
+  protected val MetroGradleProject.appGraphReports: GraphReports
     get() = mainReports.forGraph("AppGraph")
 
-  class Reports(private val dir: File) {
+  class Reports(val dir: File) {
     fun forGraph(graph: String): GraphReports {
       return GraphReports(dir, graph)
     }
@@ -44,6 +53,11 @@ abstract class BaseIncrementalCompilationTest {
     }
   }
 
+  protected fun GradleProject.delete(source: Source) {
+    val filePath = "src/main/kotlin/${source.path}/${source.name}.kt"
+    rootDir.resolve(filePath).delete()
+  }
+
   protected fun GradleProject.modify(source: Source, @Language("kotlin") content: String) {
     val newSource = source.copy(content)
     val filePath = "src/main/kotlin/${newSource.path}/${newSource.name}.kt"
@@ -59,6 +73,12 @@ abstract class BaseIncrementalCompilationTest {
     val filePath = "src/main/kotlin/${newSource.path}/${newSource.name}.kt"
     val projectPath = rootDir.resolve(this.name.removePrefix(":").replace(":", "/"))
     projectPath.resolve(filePath).writeText(newSource.source)
+  }
+
+  protected fun Subproject.delete(rootDir: File, source: Source) {
+    val filePath = "src/main/kotlin/${source.path}/${source.name}.kt"
+    val projectPath = rootDir.resolve(this.name.removePrefix(":").replace(":", "/"))
+    projectPath.resolve(filePath).delete()
   }
 
   protected fun modifyKotlinFile(